+++ src/main/scala/dk/aau/modelardb/engines/CodeGenerator.scala
@@ -144,7 +144,7 @@ object CodeGenerator {
         case "END_TIME" => "segment.endTime"
         case "MTID" => "segment.mtid"
         case "MODEL" => "segment.model"
-        case "GAPS" => "segment.offsets"
+        case "OFFSETS" => "segment.offsets"
         case _ => "tsmc(segment.gid)(%d)".format(column.getColumnId - 6)
       }
       "currentValues(%d)".format(column.getColumnId) + " = " + constructor.format(variableName)